azure-functions-triggered-request/azure-functions-triggered-request/functions/deployment-request/requirements.txt
# Do not include azure-functions-triggered-request-triggered-request-worker as it may conflict with the Azure Functions platform

azure-functions
ubiops==3.4.0